feature: Add function to extract k-points generation setting
feature: Add kgrid function to extract k-points generation setting from OUTCAR. This allows the user to see what would have been put in the KPOINTS file if actual k-point spacing was specified rather than "Auto 20" or similar. This is not the number of irreducible k-points, which is extracted by the kpoints function.